% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cpmNorm}
\alias{cpmNorm}
\title{Do log2CPM normalization with or without full library size}
\usage{
cpmNorm(rawcount, libsize = NULL)
}
\arguments{
\item{rawcount}{Matrix representing the raw read counts for samples. Row are the peaks, column are the samples}

\item{libsize}{Numeric vector representing the full library size for all samples.
If it is null, use the sum of read count in each sample as the library size.
By default, user should use full library size to scale the read counts}
}
\value{
normalized read counts \code{(log2(read counts / sum of read counts + 1))} .
}
\description{
Do log2CPM normalization with or without full library size
}
\examples{
cpmD = cpmNorm(countAll)

# The robustness.
cpmD = cpmNorm(countAll, libsize=fd$lsIP)
cpmD = cpmNorm(countAll)
}
