% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getCommPeaks}
\alias{getCommPeaks}
\title{Do overlap over input peaks files in same condition, then merge the control and treatment peaks}
\usage{
getCommPeaks(inputF, controlName = "", treatmentName = "", peakLen = 2)
}
\arguments{
\item{inputF}{file path, the file format is same to DiffBind input file to show the sample name,
TF, Control, Treatment, bam file path, and peak files path.}

\item{controlName}{string The control name}

\item{treatmentName}{string The treatment name}

\item{peakLen}{integer The length of peak threshold (>2)}
}
\value{
list
        pr GRanges objects, common peak coordinates between two conditions
        peakL list of GRanges objects, which contains the peak coordinate in each peak file
}
\description{
Do overlap over input peaks files in same condition, then merge the control and treatment peaks
}
\examples{
peaksAll = getCommPeaks(inputF, controlName="", treatmentName="",)

}
