% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getLibsize}
\alias{getLibsize}
\title{Get read counts for each sample from the bam files.}
\usage{
getLibsize(SampleID, bamFils, nCore = 1)
}
\arguments{
\item{SampleID}{Matrix representing the raw read counts for samples. Row are the peaks, column are the samples}

\item{lib.size}{Numeric vector representing the full library size for all samples.
If it is null, use the sum of read count in each sample as the library size.}

\item{group}{vector or factor giving the experimental group/condition for each sample/library.
If it is null,no control/condition are used.}
}
\value{
normalized read counts \code{(log2(read counts / sum of read counts + 1))} .
}
\description{
Get read counts for each sample from the bam files.
}
\examples{
cpmD = cpmNorm(countAll)

# The robustness.
cpmD = cpmNorm(countAll, libsize=fd$lsIP)
cpmD = cpmNorm(countAll)
}
