% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPeakCounts}
\alias{getPeakCounts}
\title{Get read counts from IP and Input bam file with given peak regions. To remove the backgroup noise,
compute the scale factor between the IP and input read counts from the non-peak regions
by by (All IP reads - reads of IP peaks regions ) / (All Input reads - reads of Input peaks regions)
The final read count of peak region are
reads of IP peak regions - scale factor * reads of input peak region}
\usage{
getPeakCounts(
  pr,
  ps,
  bamIP,
  bamInput = NULL,
  removedup = TRUE,
  fragment = 300,
  scaleControl = TRUE,
  libsize = c(1e+06, 1e+06),
  removeBackground = TRUE,
  paired.end = c("ignore", "filter")
)
}
\arguments{
\item{pr}{GRanges objects, common or merged peak coordinates between two conditions}

\item{ps}{GRanges objects, raw peak coordinates for the IP sample, if it is null}

\item{bamIP}{string IP bam file path}

\item{bamInput}{string Input bam file path, if it is null, do scale based on IP only}

\item{removedup}{logical default is true to remove duplicates reads.}

\item{fragment}{integer default is NULL, the fragment length will be determined by SGSeq::getBamInfo().
User can set it as 0, to avoid shift the reads.}

\item{scaleControl}{logical default is TRUE, do scale for read counts}

\item{removeBackground}{logical default is TRUE, remove background read count from each sample;
For IP and Input, the IP will}

\item{paired.end}{c("ignore", "filter"), refer to bamsignals-methods {bamsignals}
ignore is used for single end reads, filter is used for paired end reads
If paired.end="filter" then only first reads in proper mapped pairs will be
considered (SAMFLAG 66, i.e. in the flag of the read, the bits in the mask 66 must be all ones).}

\item{libszie}{integer vector corresponding to IP and Input libsize.}
}
\value{
list
        - count integer vector for each peak region
        - peakPos GRanges objects, common peak coordinates between two conditions
        - cIP integer read counts of IP sample
        - cInput integer read counts of input sample
}
\description{
Get read counts from IP and Input bam file with given peak regions. To remove the backgroup noise,
compute the scale factor between the IP and input read counts from the non-peak regions
by by (All IP reads - reads of IP peaks regions ) / (All Input reads - reads of Input peaks regions)
The final read count of peak region are
reads of IP peak regions - scale factor * reads of input peak region
}
\examples{
readL = getPeakCounts(pr, bamIP, bamInput,removedup=TRUE, fragment = 300,
                         scaleControl=TRUE, libsize=c(1e6,1e6))

}
