% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPeakRegion}
\alias{getPeakRegion}
\title{Do overlap over input peaks files in same condition, then merge the control and treatment peaks}
\usage{
getPeakRegion(fd)
}
\arguments{
\item{fd}{dataframe the file format is  input file to show the sample name,
TF, Control, Treatment, bam file path, and peak files path.}
}
\value{
list
        peakL list of GRanges objects, peak coordinates in each peak file
        po GRanges objects, which contains overlapped peak coordinate in each peak file
        pu GRanges objects, which contains union peak coordinate in each peak file
}
\description{
Do overlap over input peaks files in same condition, then merge the control and treatment peaks
}
\examples{
peaksAll = getPeakRegion(fd)

}
