% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getVB}
\alias{getVB}
\title{Use ridge regression to shrink the variance}
\usage{
getVB(count, fit0, design0)
}
\arguments{
\item{count}{matrix, normalized read count, row-peaks, col-samples}

\item{fit0}{object from lmFit}

\item{design}{integer The length of peak threshold (>2)}
}
\value{
list
        - coefB matrix, coefficient of ridge regression
        - varBeta matrix, residual variance of coefficient
        - lambda vector, the lambda used in ridge regression
}
\description{
Use ridge regression to shrink the variance
}
\examples{
bv = getVB(count, fit0, design)

}
