% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mrnFactors}
\alias{mrnFactors}
\title{Use MRN with or without full library size}
\usage{
mrnFactors(rawCounts, conditions, libsize = NULL)
}
\arguments{
\item{conditions}{vector or factor giving the experimental group/condition for each sample/library.
If it is null,no control/condition are used.}

\item{rawcount}{Matrix representing the raw read counts for samples. Row are the peaks, column are the samples}

\item{lib.size}{Numeric vector representing the full library size for all samples.
If it is null, use the sum of read count in each sample as the library size.}
}
\value{
normalized read counts \code{(log2(read counts / sum of read counts + 1))} .
}
\description{
Use MRN with or without full library size
}
\examples{
cpmD = mrnFactors(countAll,conditions=c(rep(1,2), rep(2,2)), libsize=fd$lsIP)

}
