% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{omPeaks}
\alias{omPeaks}
\title{Do merge or overlap over input peaks files based their coordinates}
\usage{
omPeaks(peakFiles, overlapOrMerge = TRUE, len = 2)
}
\arguments{
\item{peakFiles}{path vector of input peak files}

\item{overlapOrMerge}{logical If TURE, do overlap; else do merge}

\item{len}{numeric, peak length to filter peaks}
}
\value{
GRanges objects, merged or ovelapped peak coordinates.
}
\description{
Do merge or overlap over input peaks files based their coordinates
}
\examples{
peaksO = omPeaks(peakID, overlapOrMerge=TRUE)

}
