% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotCorBoxs}
\alias{plotCorBoxs}
\title{Plot Box plot of normalized samples}
\usage{
plotCorBoxs(df, title = "", xn = "Samples", yn = "log2(CPM+1)")
}
\arguments{
\item{df}{input matrix with colnames as samples}

\item{title}{string}

\item{xn}{name of x-axis}

\item{yn}{name of y-axis}
}
\value{
p2 ggplot
}
\description{
Plot Box plot of normalized samples
}
\examples{
plotCorBoxs(count)

}
