% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotMAVoc}
\alias{plotMAVoc}
\title{Plot MA plot or volcano plot}
\usage{
plotMAVoc(
  mean,
  logfc,
  adj.P.Val,
  refDE = NULL,
  FC = 0,
  padjT = 0.05,
  psize = 1,
  title = "",
  MA = FALSE
)
}
\arguments{
\item{mean}{vector}

\item{logfc}{vector}

\item{adj.P.Val}{vector}

\item{refDE}{vector known real DE annotation: raw, Up, and Down}

\item{FC}{numeric, draw logFC line with given FC value}

\item{psize}{numeric, adjust the point size}

\item{title}{string, the title of plot}

\item{MA}{logical, if is true, MA plot. else, volcano plot}

\item{padj}{numeric, filter the DE peaks with padj vlaue}
}
\value{
p2 ggplot
}
\description{
Plot MA plot or volcano plot
}
\examples{
bv = getVB(count, fit0, design)

}
