% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ridgeReg}
\alias{ridgeReg}
\title{Use ridge regression to fit the read count for each region given the design}
\usage{
ridgeReg(oy, Xd, lambda = 1)
}
\arguments{
\item{oy}{vector, the read counts for all samples}

\item{Xd}{design matrix, is same to limma's design}

\item{lambda}{numeric, used as lambda for ridge regression (0<= lambda <=1)}
}
\value{
list
        - betaRlm vector, coefficient of ridge regression
        - varBeta vector, residual variance of coefficient
        - mse numeric, the MSE of ridge regression
}
\description{
Use ridge regression to fit the read count for each region given the design
}
\examples{
rg = ridgeReg(oy,Xd, lambda=0.01)

}
